Add looping, menu input, var init, struct
Adds feature to loop until the user gives the option to quit. Adds
the feature for users to enter an option.  Invalid options cause
exit(-1). Adds support for the LIST_ALL command and QUIT command.
Created a struct for our entries.  Also included a typedef and the
global variable to contain all of the entries.

First time successfully using tags.
1. run ctags -R
2. in vim, run :load tags
3. use Ctrl-] to follow tags, Ctrl-t to go back up the stack