Tried Compile-time asserts.  Worked on add functionality.
This commit does not compile.  Unfortunately neither did the previous
one.  I didn't realize that the add functionality wasn't finished.
I worked on it this time around but it's still not done.

Tried to implement a hacked-up compile time assert because only very
recent versions of the C-language support it.  Was hard to tell if it
was working or not due to the unfinished nature of the add function.