Got program compiling with add function, but there are bugs
Got the program compiling with the add function, however it's not
acting the way that it's supposed to.  I'm not sure if this is because
of the add function, the list-all function, or something else that i'm
doing.  It's possible that the memory isn't getting copied correctly
to a global location.  I may have to dig in with gdb to see where the
data is going.

Also added another header file to hold utility macros.  Currently it
just has a MIN and a MAX macro.