Implement VIEW ENTRY command
Implemented the command to view an individual entry.  However, it
doesn't seem to work.  It seems to always give you the first entry
regardless.

I'm not sure if there's a problem with the function that searches
for the names, or the one that tries to check if it's a valid id
first.

Also added LOAD FILE to the list of commands, but didn't implement
it yet.