Fixed detecting "enter" key. Fixed name search.
Fixed the problem of detecting when empty text is submitted with the
enter key.  It involves using the function fgets.  However, there
is a problem with interchanging fgets and scanf.  But it seems to
work out in this case, b/c it just prints the menu an extra time
and then continues to work normally.

After debugging with gdb, it was discovered that strncpy was not copying
over the null character at the end of the string.  To fix the problem,
i used a memset before hand to zero out all the characters in the array.
However, now that i'm thinking about it, i could probably modify the
copy to grab 1 extra character.